Cumulative changes (#208)
* Bump jedis from 3.7.1 to 4.1.1 (#52)

* dependency resolution
* fix tests


* Bump maven-compiler-plugin from 3.9.0 to 3.10.1 (#63)

Bumps [maven-compiler-plugin](https://github.com/apache/maven-compiler-plugin) from 3.9.0 to 3.10.1.
- [Release notes](https://github.com/apache/maven-compiler-plugin/releases)
- [Commits](https://github.com/apache/maven-compiler-plugin/compare/maven-compiler-plugin-3.9.0...maven-compiler-plugin-3.10.1)

---
updated-dependencies:
- dependency-name: org.apache.maven.plugins:maven-compiler-plugin
  dependency-type: direct:production
  update-type: version-update:semver-minor
...



* Bump mockito-core from 4.3.0 to 4.4.0 (#62)

Bumps [mockito-core](https://github.com/mockito/mockito) from 4.3.0 to 4.4.0.
- [Release notes](https://github.com/mockito/mockito/releases)
- [Commits](https://github.com/mockito/mockito/compare/v4.3.0...v4.4.0)

---
updated-dependencies:
- dependency-name: org.mockito:mockito-core
  dependency-type: direct:development
  update-type: version-update:semver-minor
...



* Bump maven-pmd-plugin from 3.15.0 to 3.16.0 (#57)

Bumps [maven-pmd-plugin](https://github.com/apache/maven-pmd-plugin) from 3.15.0 to 3.16.0.
- [Release notes](https://github.com/apache/maven-pmd-plugin/releases)
- [Commits](https://github.com/apache/maven-pmd-plugin/compare/maven-pmd-plugin-3.15.0...maven-pmd-plugin-3.16.0)

---
updated-dependencies:
- dependency-name: org.apache.maven.plugins:maven-pmd-plugin
  dependency-type: direct:production
  update-type: version-update:semver-minor
...



* Bump maven-javadoc-plugin from 3.3.1 to 3.3.2 (#54)

Bumps [maven-javadoc-plugin](https://github.com/apache/maven-javadoc-plugin) from 3.3.1 to 3.3.2.
- [Release notes](https://github.com/apache/maven-javadoc-plugin/releases)
- [Commits](https://github.com/apache/maven-javadoc-plugin/compare/maven-javadoc-plugin-3.3.1...maven-javadoc-plugin-3.3.2)

---
updated-dependencies:
- dependency-name: org.apache.maven.plugins:maven-javadoc-plugin
  dependency-type: direct:production
  update-type: version-update:semver-patch
...



* Bump mockito-core from 4.0.0 to 4.1.0 (#24) (#53)

Bumps [mockito-core](https://github.com/mockito/mockito) from 4.0.0 to 4.1.0.
- [Release notes](https://github.com/mockito/mockito/releases)
- [Commits](https://github.com/mockito/mockito/compare/v4.0.0...v4.1.0)

---
updated-dependencies:
- dependency-name: org.mockito:mockito-core
  dependency-type: direct:development
  update-type: version-update:semver-minor
...




* zrange/zrevrange support (#59)

* zrevrange support

* fix ZRange and add REV support


* Update list of supported operations

* Fix 'HSET after EXPIRE discards the expiration' (#79)

* fixed issue 187 and added test


* Fix 'HKEYS/HLEN/HVALS/HGETALL fails when key does not exist' (#78)

* fixed issue 191

* Fix 'SETBIT clears other bits' (#77)

* fixed issue 189

* Bump spotbugs-maven-plugin from 4.5.3.0 to 4.6.0.0 (#75)

Bumps [spotbugs-maven-plugin](https://github.com/spotbugs/spotbugs-maven-plugin) from 4.5.3.0 to 4.6.0.0.
- [Release notes](https://github.com/spotbugs/spotbugs-maven-plugin/releases)
- [Commits](https://github.com/spotbugs/spotbugs-maven-plugin/compare/spotbugs-maven-plugin-4.5.3.0...spotbugs-maven-plugin-4.6.0.0)

---
updated-dependencies:
- dependency-name: com.github.spotbugs:spotbugs-maven-plugin
  dependency-type: direct:production
  update-type: version-update:semver-minor
...



* Bump jedis from 4.1.1 to 4.2.0 (#74)

Bumps [jedis](https://github.com/redis/jedis) from 4.1.1 to 4.2.0.
- [Release notes](https://github.com/redis/jedis/releases)
- [Commits](https://github.com/redis/jedis/compare/v4.1.1...v4.2.0)

---
updated-dependencies:
- dependency-name: redis.clients:jedis
  dependency-type: direct:development
  update-type: version-update:semver-minor
...



* Bump lettuce-core from 6.1.6.RELEASE to 6.1.8.RELEASE (#64)

Bumps [lettuce-core](https://github.com/lettuce-io/lettuce-core) from 6.1.6.RELEASE to 6.1.8.RELEASE.
- [Release notes](https://github.com/lettuce-io/lettuce-core/releases)
- [Changelog](https://github.com/lettuce-io/lettuce-core/blob/6.1.8.RELEASE/RELEASE-NOTES.md)
- [Commits](https://github.com/lettuce-io/lettuce-core/compare/6.1.6.RELEASE...6.1.8.RELEASE)

---
updated-dependencies:
- dependency-name: io.lettuce:lettuce-core
  dependency-type: direct:development
  update-type: version-update:semver-patch
...



* Fix 'HDEL does not work on additional arguments' (#76)

* fixed issue 192

* add test

* refactor tests


* Update ZRANGE support for 6.2.0, implement ZREVRANGE, ZREVRANGEBYSCORE (#80)

* fix ZRange

* add tests for ZRange

* Bump jedis from 4.2.0 to 4.2.1 (#81)

Bumps [jedis](https://github.com/redis/jedis) from 4.2.0 to 4.2.1.
- [Release notes](https://github.com/redis/jedis/releases)
- [Commits](https://github.com/redis/jedis/compare/v4.2.0...v4.2.1)

---
updated-dependencies:
- dependency-name: redis.clients:jedis
  dependency-type: direct:development
  update-type: version-update:semver-patch
...



* Update list of supported operations

+++ src/test/java/com/github/fppt/jedismock/comparisontests/keys/KeysOperationsTest.java
@@ -68,7 +68,7 @@ public class KeysOperationsTest {
         jedis.hset(key2, "field", "value");
         assertTrue(jedis.exists(key1));
         assertTrue(jedis.exists(key2));
-        int count = jedis.del(key1, key2).intValue();
+        long count = jedis.del(key1, key2);
         assertEquals(2, count);
         assertFalse(jedis.exists(key1));
         assertFalse(jedis.exists(key2));
